time is not int, not Fraction

+++ src/main/java/ptcs/graphics/GraphicsUtils.java
@@ -18,8 +18,8 @@ public class GraphicsUtils {
         double scaleMax, Color c) {
 
         g.setColor(c);
-        double x1 = MyMath.scale_3_2(scaleMin, n.tOn.doubleValue(), scaleMax, 0, width);
-        double x2 = MyMath.scale_3_2(scaleMin, n.tOff.doubleValue(), scaleMax, 0, width);
+        double x1 = MyMath.scale_3_2(scaleMin, n.tOn, scaleMax, 0, width);
+        double x2 = MyMath.scale_3_2(scaleMin, n.tOff, scaleMax, 0, width);
         double y1 = MyMath.scale_3_2(20, n.pitch, 90, height, 0);
         g.drawLine((int)x1, (int)y1, (int)x2, (int)y1);
